+++ dataforge-meta/src/commonTest/kotlin/space/kscience/dataforge/meta/MutableMetaTest.kt
@@ -3,9 +3,9 @@ package space.kscience.dataforge.meta
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class MutableMetaTest{
+class MutableMetaTest {
     @Test
-    fun remove(){
+    fun remove() {
         val meta = MutableMeta {
             "aNode" put {
                 "innerNode" put {
@@ -18,5 +18,27 @@ class MutableMetaTest{
 
         meta.remove("aNode.c")
         assertEquals(meta["aNode.c"], null)
+    }
+
+    @Test
+    fun withDefault() {
+        val metaWithDefault = MutableMeta().withDefault(
+            Meta {
+                "a" put {
+                    value = 22.asValue()
+                    "b" put true
+                }
+            }
+        )
+
+        assertEquals(22, metaWithDefault["a"].int)
+        assertEquals(true, metaWithDefault.getValue("a.b")?.boolean)
+
+        metaWithDefault["a.b"] = "false"
+
+        assertEquals(false, metaWithDefault["a.b"]?.boolean)
+        assertEquals(false, metaWithDefault.getValue("a.b")?.boolean)
+        assertEquals(22, metaWithDefault.getValue("a")?.int)
+
     }
 }